refactor:module despesas

+++ app/Services/Usuarios/DeletarUsuarioService.php
@@ -10,6 +10,13 @@ class DeletarUsuarioService
 {
     public function handler(array $options)
     {
+        if (auth()->payload()->get('sub') != Arr::get($options, 'id')) {
+            throw new HttpResponseException(response()->json([
+                'message' => 'Erro, Você não tem permissão.',
+                'success' => false
+            ], 403, ['Content-Type' => 'application/json;charset=UTF-8', 'Charset' => 'utf-8'], JSON_UNESCAPED_UNICODE));
+        }
+
         try {
             $usuarioRepo = new UsuarioRepo();
             $usuarioRepo->delete($options);
